refactor(canton): extract stream.rs from indexer_canton
